Enrich and cache authoritative showtime labels
