+++ .github/workflows/deploy.yml
@@ -28,8 +28,22 @@ jobs:
 
           echo "NODE_EXTRA_CA_CERTS=$RUNNER_TEMP/kvikmyndir-chain.pem" >> "$GITHUB_ENV"
 
+      - name: ♻️ Restore showtime metadata cache
+        uses: actions/cache/restore@main
+        with:
+          path: static/showtime-metadata-cache.json
+          key: showtime-metadata-${{ github.run_id }}
+          restore-keys: |
+            showtime-metadata-
+
       - name: 🎬 Scrape movie data
         run: bun scrape
+
+      - name: 💾 Save showtime metadata cache
+        uses: actions/cache/save@main
+        with:
+          path: static/showtime-metadata-cache.json
+          key: showtime-metadata-${{ github.run_id }}
 
       - name: ☑️ Check svelte-kit, format and linting
         run: bun check
