+++ src/lib/CinemaShowtimeRow.svelte
@@ -40,7 +40,7 @@
         <span>{new Date(showtime.time).toLocaleTimeString("is-IS", { timeStyle: "short", hour12: false })}</span>
         <span
           class="pointer-events-none absolute bottom-full left-1/2 mb-1.5 hidden -translate-x-1/2 rounded bg-neutral-950/95 px-2 py-1 text-[10px] font-medium whitespace-nowrap text-neutral-300 opacity-0 shadow-lg transition-opacity group-hover/time:opacity-100 [@media(hover:hover)]:block">
-          Kaupa miða
+          {showtime.hall ? `${showtime.hall} · ` : ""}Kaupa miða
         </span>
       </a>
     {/each}
